feat: simply client interface and rename client to deepseek_client to avoud conflict with other package
