+++ deepseek-api/src/client_builder.rs
@@ -6,31 +6,31 @@ cfg_if::cfg_if! {
         use reqwest::ClientBuilder as ReqwestClientBuilder;
     }
 }
-use crate::Client;
+use crate::DeepSeekClient;
 use reqwest::header::HeaderMap;
 use std::time::Duration;
 
-/// A builder for constructing a `Client` instance with customizable options.
+/// A builder for constructing a `DeepSeekClient` instance with customizable options.
 ///
-/// The `ClientBuilder` allows you to configure the API key, timeout, and host
-/// for the `Client` before building it.
+/// The `DeepSeekClientBuilder` allows you to configure the API key, timeout, and host
+/// for the `DeepSeekClient` before building it.
 ///
 /// # Examples
 ///
 /// ```ignore
-/// let client = ClientBuilder::new("your_api_key".to_string())
+/// let client = DeepSeekClientBuilder::new("your_api_key".to_string())
 ///     .timeout(30)
 ///     .build()
 ///     .expect("Failed to build client");
 /// ```
-pub struct ClientBuilder {
+pub struct DeepSeekClientBuilder {
     api_key: String,
     timeout: Option<u64>,
     host: String,
 }
 
-impl ClientBuilder {
-    /// Creates a new `ClientBuilder` with the specified API key.
+impl DeepSeekClientBuilder {
+    /// Creates a new `DeepSeekClientBuilder` with the specified API key.
     ///
     /// # Arguments
     ///
@@ -38,7 +38,7 @@ impl ClientBuilder {
     ///
     /// # Returns
     ///
-    /// A new instance of `ClientBuilder` with default settings.
+    /// A new instance of `DeepSeekClientBuilder` with default settings.
     ///
     /// The default host is set to `"https://api.deepseek.com"`, and no timeout is configured.
     pub fn new(api_key: String) -> Self {
@@ -57,9 +57,9 @@ impl ClientBuilder {
     ///
     /// # Returns
     ///
-    /// The `ClientBuilder` instance with the timeout configured.
+    /// The `DeepSeekClientBuilder` instance with the timeout configured.
     /// ```ignore
-    /// let builder = ClientBuilder::new("your_api_key".to_string())
+    /// let builder = DeepSeekClientBuilder::new("your_api_key".to_string())
     ///     .timeout(30);
     /// ```
     pub fn timeout(mut self, duration: u64) -> Self {
@@ -76,18 +76,18 @@ impl ClientBuilder {
     ///
     /// # Errors
     ///
-    /// This method will return an error if the underlying `reqwest::blocking::ClientBuilder`
+    /// This method will return an error if the underlying `reqwest::blocking::DeepSeekClientBuilder`
     /// fails to build the client.
     ///
     /// # Examples
     ///
     /// ```ignore
-    /// let client = ClientBuilder::new("your_api_key".to_string())
+    /// let client = DeepSeekClientBuilder::new("your_api_key".to_string())
     ///     .timeout(30)
     ///     .build()
     ///     .expect("Failed to build client");
     /// ```
-    pub fn build(self) -> Result<Client> {
+    pub fn build(self) -> Result<DeepSeekClient> {
         let mut headers = HeaderMap::new();
         headers.insert("Authorization", format!("Bearer {}", self.api_key).parse()?);
 
@@ -99,7 +99,7 @@ impl ClientBuilder {
         };
 
         let client = client_builder.build()?;
-        Ok(Client {
+        Ok(DeepSeekClient {
             client,
             host: self.host,
         })
