+++ deepseek-api/src/lib.rs
@@ -13,12 +13,10 @@ cfg_if::cfg_if! {
 
         pub use sync_impl::json_stream;
         pub use sync_impl::client::*;
-        pub use sync_impl::completions;
     } else {
         mod async_impl;
 
         pub use async_impl::json_stream;
         pub use async_impl::client::*;
-        pub use async_impl::completions;
     }
 }
