+++ deepseek-api/src/sync_impl/client.rs
@@ -1,5 +1,9 @@
-use super::completions::ChatCompletions;
-use crate::response::{BalanceResp, ModelResp};
+use super::error::ToApiError;
+use super::json_stream::JsonStream;
+use crate::{
+    response::{BalanceResp, ChatResponse, ModelResp},
+    RequestBuilder,
+};
 use anyhow::Result;
 use reqwest::blocking::Client as ReqwestClient;
 
@@ -9,38 +13,28 @@ use reqwest::blocking::Client as ReqwestClient;
 /// # Example
 ///
 /// ```no_run
-///     use deepseek_api::ClientBuilder;
+///     use deepseek_api::DeepSeekClientBuilder;
 ///
 ///     let api_key = "your_api_key".to_string();
-///     let client = ClientBuilder::new(api_key).build().unwrap();
+///     let client = DeepSeekClientBuilder::new(api_key).build().unwrap();
 ///
 ///     // Get available models
 ///     let models = client.models().unwrap();
 ///
 ///     // Get user balance
 ///     let balance = client.balance().unwrap();
-///
-///     // Create a chat completion
-///     let chat = client.chat();
 /// ```
 ///
 /// # Fields
 ///
 /// * `client` - The underlying HTTP client.
 /// * `host` - The base URL for the DeepSeek API.
-pub struct Client {
+pub struct DeepSeekClient {
     pub(crate) client: ReqwestClient,
     pub(crate) host: String,
 }
 
-impl Client {
-    pub fn chat(&self) -> ChatCompletions {
-        ChatCompletions {
-            client: self.client.clone(),
-            host: self.host.clone(),
-        }
-    }
-
+impl DeepSeekClient {
     /// Retrieves the list of available models from the DeepSeek API.
     ///
     /// This method sends a GET request to the `/models` endpoint of the DeepSeek API
@@ -54,10 +48,10 @@ impl Client {
     /// # Example
     ///
     /// ```no_run
-    ///     use deepseek_api::ClientBuilder;
+    ///     use deepseek_api::DeepSeekClientBuilder;
     ///
     ///     let api_key = "your_api_key".to_string();
-    ///     let client = ClientBuilder::new(api_key).build().unwrap();
+    ///     let client = DeepSeekClientBuilder::new(api_key).build().unwrap();
     ///     let models = client.models().unwrap();
     ///     println!("{:?}", models);
     /// ```
@@ -84,10 +78,10 @@ impl Client {
     /// # Example
     ///
     /// ```no_run
-    ///     use deepseek_api::ClientBuilder;
+    ///     use deepseek_api::DeepSeekClientBuilder;
     ///
     ///     let api_key = "your_api_key".to_string();
-    ///     let client = ClientBuilder::new(api_key).build().unwrap();
+    ///     let client = DeepSeekClientBuilder::new(api_key).build().unwrap();
     ///     let balance = client.balance().unwrap();
     ///     println!("{:?}", balance);
     /// ```
@@ -99,5 +93,89 @@ impl Client {
             .get(self.host.to_owned() + "/user/balance")
             .send()?
             .json()?)
+    }
+
+    /// Sends a completion request to the DeepSeek API.
+    ///
+    /// This method constructs a request using the provided `RequestBuilder` and sends it
+    /// to the appropriate endpoint of the DeepSeek API. Depending on whether the request
+    /// is for a beta feature or not, it will target either the `/beta/completions` or
+    /// `/chat/completions` endpoint. The response can be either a full response or a
+    /// streaming response, based on the `stream` optional of the `RequestBuilder`.
+    ///
+    /// # Type Parameters
+    ///
+    /// * `Builder` - A type that implements the `RequestBuilder` trait, used to construct
+    ///   the request payload.
+    ///
+    /// # Arguments
+    ///
+    /// * `request_builder` - An instance of a type implementing the `RequestBuilder` trait,
+    ///   which is used to build the request payload.
+    ///
+    /// # Returns
+    ///
+    /// A `Result` containing a `ChatResponse` on success. The `ChatResponse` can either be
+    /// a full response or a streaming response, depending on the request.
+    ///
+    /// # Errors
+    ///
+    /// This function will return an error if:
+    /// - The request fails to send.
+    /// - The response contains an API error.
+    /// - The response cannot be deserialized into the expected type.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    ///     use deepseek_api::{request::{MessageRequest, UserMessageRequest}, DeepSeekClientBuilder, CompletionsRequestBuilder};
+    ///     use deepseek_api::response::ChatResponse;
+    ///     use futures_util::StreamExt;
+    ///
+    ///     let api_key = "your_api_key".to_string();
+    ///     let client = DeepSeekClientBuilder::new(api_key).build().unwrap();
+    ///     let request_builder = CompletionsRequestBuilder::new(vec![MessageRequest::User(
+    ///                 UserMessageRequest::new("Hello, DeepSeek!")
+    ///     )]);
+    ///
+    ///     let response = client.send_completion_request(request_builder).unwrap();
+    ///     match response {
+    ///         ChatResponse::Full(full_response) => println!("{:?}", full_response),
+    ///         ChatResponse::Stream(mut stream) => {
+    ///             while let Some(item) = stream.next() {
+    ///                 println!("{:?}", item);
+    ///             }
+    ///         }
+    ///     }
+    /// ```
+    ///
+    /// For more information, see the [DeepSeek API documentation](https://api-docs.deepseek.com/zh-cn/api).
+    pub fn send_completion_request<Builder>(
+        &self,
+        request_builder: Builder,
+    ) -> Result<ChatResponse<Builder::Response, Builder::Item>>
+    where
+        Builder: RequestBuilder + Send + Sized,
+    {
+        let host = if request_builder.is_beta() {
+            self.host.to_owned() + "/beta/completions"
+        } else {
+            self.host.to_owned() + "/chat/completions"
+        };
+        let is_stream = request_builder.is_stream();
+
+        let request = request_builder.build();
+        let resp = self
+            .client
+            .post(&host)
+            .json(&request)
+            .send()?
+            .to_api_err()?;
+
+        if is_stream {
+            Ok(ChatResponse::Stream(JsonStream::new(resp)))
+        } else {
+            Ok(ChatResponse::Full(resp.json()?))
+        }
     }
 }
