--- deepseek-api/src/sync_impl/completions.rs
@@ -1,41 +1,0 @@
-use super::json_stream::JsonStream;
-use crate::{response::ChatResponse, RequestBuilder};
-use anyhow::Result;
-use reqwest::blocking::Client as ReqwestClient;
-
-use super::error::ToApiError;
-pub struct ChatCompletions {
-    pub(crate) client: ReqwestClient,
-    pub(crate) host: String,
-}
-
-impl ChatCompletions {
-    pub fn create<Builder>(
-        &self,
-        request_builder: Builder,
-    ) -> Result<ChatResponse<Builder::Response, Builder::Item>>
-    where
-        Builder: RequestBuilder + Send + Sized,
-    {
-        let host = if request_builder.is_beta() {
-            self.host.to_owned() + "/beta/completions"
-        } else {
-            self.host.to_owned() + "/chat/completions"
-        };
-        let is_stream = request_builder.is_stream();
-
-        let request = request_builder.build();
-        let resp = self
-            .client
-            .post(&host)
-            .json(&request)
-            .send()?
-            .to_api_err()?;
-
-        if is_stream {
-            Ok(ChatResponse::Stream(JsonStream::new(resp)))
-        } else {
-            Ok(ChatResponse::Full(resp.json()?))
-        }
-    }
-}
