+++ deepseek-api/src/sync_impl/mod.rs
@@ -1,4 +1,3 @@
 pub mod client;
-pub mod completions;
 mod error;
 pub mod json_stream;
