+++ examples/chat-stream/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use clap::Parser;
-use deepseek_api::{ClientBuilder, FMICompletionsRequestBuilder, RequestBuilder};
+use deepseek_api::{DeepSeekClientBuilder, FMICompletionsRequestBuilder, RequestBuilder};
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -12,7 +12,7 @@ struct Args {
 fn main() -> Result<()> {
     let args = Args::parse();
 
-    let client = ClientBuilder::new(args.api_key.clone()).build()?;
+    let client = DeepSeekClientBuilder::new(args.api_key.clone()).build()?;
 
     let balances = client.balance()?;
     println!("balances {:?}", balances);
@@ -20,11 +20,10 @@ fn main() -> Result<()> {
     let models = client.models()?;
     println!("models {:?}", models);
 
-    let completions = client.chat();
     let stream = FMICompletionsRequestBuilder::new("def fib(a):", "    return fib(a-1) + fib(a-2)")
         .max_tokens(128)?
         .stream(true)
-        .do_request(&completions)?
+        .do_request(&client)?
         .must_stream();
     for item in stream {
         println!("resp: {:?}", item);
