+++ examples/fim/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use clap::Parser;
-use deepseek_api::{ClientBuilder, FMICompletionsRequestBuilder, RequestBuilder};
+use deepseek_api::{DeepSeekClientBuilder, FMICompletionsRequestBuilder, RequestBuilder};
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -12,8 +12,7 @@ struct Args {
 #[tokio::main]
 async fn main() -> Result<()> {
     let args = Args::parse();
-
-    let client = ClientBuilder::new(args.api_key.clone()).build()?;
+    let client = DeepSeekClientBuilder::new(args.api_key.clone()).build()?;
 
     let balances = client.balance().await?;
     println!("balances {:?}", balances);
@@ -21,10 +20,9 @@ async fn main() -> Result<()> {
     let models = client.models().await?;
     println!("models {:?}", models);
 
-    let completions = client.chat();
     let resp = FMICompletionsRequestBuilder::new("def fib(a):", "    return fib(a-1) + fib(a-2)")
         .max_tokens(128)?
-        .do_request(&completions)
+        .do_request(&client)
         .await?
         .must_response();
     println!(
